feat(Step07): JPA 엔티티 변환 및 간접 참조 패턴 적용
- Order, OrderItem 엔티티 JPA 변환 완료
- Cart, CartItem 엔티티 JPA 변환 완료
- 엔티티에서 items 필드 제거하여 완전한 간접 참조 구현
- JPA Pageable을 사용한 페이징 성능 개선
- OrderItemRepository, CartItemRepository 추가
- 모든 Repository를 JpaRepository로 변환
- InMemory 구현체 제거
- ERD에 carts 테이블 추가

+++ src/main/java/com/sparta/ecommerce/application/coupon/dto/CouponResponse.java
@@ -1,6 +1,6 @@
 package com.sparta.ecommerce.application.coupon.dto;
 
-import com.sparta.ecommerce.domain.coupon.Coupon;
+import com.sparta.ecommerce.domain.coupon.entity.Coupon;
 import com.sparta.ecommerce.domain.coupon.DiscountType;
 import io.swagger.v3.oas.annotations.media.Schema;
 
@@ -20,7 +20,7 @@ public record CouponResponse(
         DiscountType discountType,
 
         @Schema(description = "할인 값 (FIXED: 금액, PERCENT: 퍼센트)", example = "50000")
-        int discountValue,
+        long discountValue,
 
         @Schema(description = "총 발급 수량", example = "100")
         int totalQuantity,
@@ -29,7 +29,7 @@ public record CouponResponse(
         int remainingQuantity,
 
         @Schema(description = "최소 주문 금액", example = "100000")
-        int minOrderAmount,
+        long minOrderAmount,
 
         @Schema(description = "만료 일시", example = "2025-12-31T23:59:59")
         LocalDateTime expiresAt
